move partition to own file
